+++ install.py
@@ -11,7 +11,7 @@ def loader():
 class BasicInstaller(ExtensionInstaller):
     def __init__(self):
         super(BasicInstaller, self).__init__(
-            version="1.3.0",
+            version="2.0.0",
             name="weewx-wdc",
             description="Weather Data Center skin for weewx.",
             author="David Baetge",
@@ -70,9 +70,12 @@ class BasicInstaller(ExtensionInstaller):
                         "skins/weewx-wdc/includes/data-table-tile.inc",
                         "skins/weewx-wdc/includes/diagram-tile.inc",
                         "skins/weewx-wdc/includes/stat-tile.inc",
+                        "skins/weewx-wdc/includes/conditions-table.inc",
                         "skins/weewx-wdc/includes/climatological-days.inc",
                         "skins/weewx-wdc/includes/ui-shell.inc",
                         "skins/weewx-wdc/includes/footer.inc",
+                        "skins/weewx-wdc/includes/body-classic.inc",
+                        "skins/weewx-wdc/includes/body-alternative.inc",
                         "skins/weewx-wdc/includes/forecast.inc",
                         "skins/weewx-wdc/includes/forecast-table.inc",
                         "skins/weewx-wdc/includes/icons/barometer.svg",